refactor(surfer): move to internal/surfer
