+++ internal/sidekick/config/config.go
@@ -22,8 +22,8 @@ import (
 	"os"
 	"path"
 
-	"github.com/googleapis/librarian/internal/sidekick/config/gcloudyaml"
 	"github.com/googleapis/librarian/internal/sidekick/license"
+	"github.com/googleapis/librarian/internal/surfer/gcloudyaml"
 	toml "github.com/pelletier/go-toml/v2"
 )
