+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/_partials/_cancel_ga.yaml
