+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/_partials/_create_ga.yaml
