+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/_partials/_delete_ga.yaml
