+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/_partials/_describe_ga.yaml
