+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/_partials/_wait_ga.yaml
