+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/create.yaml
