+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/delete.yaml
