+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/describe.yaml
