+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/export_data.yaml
