+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/list.yaml
