+++ internal/surfer/gcloud/testdata/parallelstore/surface/instances/update.yaml
