+++ internal/surfer/gcloud/testdata/parallelstore/surface/locations/describe.yaml
