+++ internal/surfer/gcloud/testdata/parallelstore/surface/locations/list.yaml
