+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/_partials/_cancel_ga.yaml
