+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/_partials/_deleta_ga.yaml
