+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/_partials/_describe_ga.yaml
