+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/_partials/_list_ga.yaml
