+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/_partials/_wait_ga.yaml
