+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/cancel.yaml
