+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/delete.yaml
