+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/describe.yaml
