+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/list.yaml
