+++ internal/surfer/gcloud/testdata/parallelstore/surface/operations/wait.yaml
