+++ internal/surfer/gcloudyaml/testdata/gcloud.yaml
